feat(core): Changed templating structure and export class
Now by default only EmbedJS class is exported and all the earlier 'ejs' methods are now static
methods of that class. This is done to make it easier to create umd build and also there is a new
Template constructor. So that users can use use different instances of templates on the same page.

BREAKING